requirements.txt
fastapi
python-dotenv
firebase-admin
google-cloud-storage
google-cloud-bigquery
uvicorn[standard]
